+++ resources/views/back/managers/_partials/context-menu.blade.php
@@ -1,8 +1,8 @@
 <options-dropdown class="inline-block">
     <div class="inset-s" v-cloak>
 
-        @if(contract($manager, \Thinktomorrow\Chief\Management\ManagerThatPreviews::class))
-            <a class="block squished-s --link-with-bg" href="{!! $manager->previewUrl() !!}" target="_blank">Bekijk preview</a>
+        @if($manager->isAssistedBy('publish'))
+            <a class="block squished-s --link-with-bg" href="{!! $manager->assistant('publish')->previewUrl() !!}" target="_blank">Bekijk preview</a>
         @endif
 
         @if($manager->can('edit') && \Illuminate\Support\Facades\Route::currentRouteName() !== 'chief.back.managers.edit')
