--- src/Management/ManagerThatPreviews.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Thinktomorrow\Chief\Management;
-
-interface ManagerThatPreviews
-{
-    public function publicationStatusAsLabel($plain = false);
-}
